requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Data handling
xarray>=2023.1.0
netCDF4>=1.6.0
h5netcdf>=1.1.0

# Image processing
Pillow>=9.5.0
matplotlib>=3.7.0
scikit-image>=0.21.0

# Machine learning utilities
scikit-learn>=1.2.0
tqdm>=4.65.0

# HTTP requests (for data download)
requests>=2.28.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.22.0
